% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance-metrics.R
\name{generate_distances}
\alias{generate_distances}
\title{Generates all distances between graphs in provided list}
\usage{
generate_distances(l, distance_function)
}
\arguments{
\item{l}{List of matricies representing graphs of interest}

\item{distance_function}{One of the distance metrics defined in this package}
}
\value{
Distance matricies
}
\description{
Generates all distances between graphs in provided list
}
