% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance-metrics.R
\name{generate_distances_mnem}
\alias{generate_distances_mnem}
\title{Generates all distances between graphs returned in an mnem object}
\usage{
generate_distances_mnem(res.list, distance_function)
}
\arguments{
\item{res.list}{As returned from mnem}

\item{distance_function}{One of the distance metrics defined in this package}
}
\value{
Vector of distance matricies
}
\description{
Generates all distances between graphs returned in an mnem object
}
