% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance-metrics.R
\name{hamming_dist}
\alias{hamming_dist}
\title{Find the Hamming distance between two graphs as sum of the absolute value differences between adjacency matricies.}
\usage{
hamming_dist(x, y)
}
\arguments{
\item{x}{An adjacency matrix}

\item{y}{An adjacency matrix}
}
\value{
Number representing the distance between x and y
}
\description{
Find the Hamming distance between two graphs as sum of the absolute value differences between adjacency matricies.
}
