\name{labnetmet-package}
\alias{labnetmet-package}
\alias{labnetmet}
\docType{package}
\title{
\packageTitle{labnetmet}
}
\description{
\packageDescription{labnetmet}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{labnetmet}
\packageIndices{labnetmet}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{labnetmet}

Maintainer: \packageMaintainer{labnetmet}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
