% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance-metrics.R
\name{plot_dist}
\alias{plot_dist}
\title{Plots the networks and a heatmap of distances between them}
\usage{
plot_dist(l, distance_function, filename = "")
}
\arguments{
\item{l}{A list of adjacency matricies representing the networks of interest}

\item{distance_function}{One of the distance metrics defined in this package}
}
\value{
Plot object
}
\description{
Plots the networks and a heatmap of distances between them
}
