% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance-metrics.R
\name{same_reachability}
\alias{same_reachability}
\title{XOR for finiteness}
\usage{
same_reachability(e, f)
}
\arguments{
\item{e}{A distance}

\item{f}{A distance}
}
\value{
True if both are either finite or infinte, False if they are not
}
\description{
XOR for finiteness
}
