% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance-metrics.R
\name{subgraph_dist}
\alias{subgraph_dist}
\title{Find the subgraph distance between two graphs as the number of subgraphs that differ between the networks.}
\usage{
subgraph_dist(x, y)
}
\arguments{
\item{x}{An adjacency matrix}

\item{y}{An adjacency matrix}
}
\value{
Number representing the distance between x and y
}
\description{
Find the subgraph distance between two graphs as the number of subgraphs that differ between the networks.
}
