% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance-metrics.R
\name{trans_dist}
\alias{trans_dist}
\title{Find the transitive distance between two graphs as the number of directed paths that differ between graphs x and y}
\usage{
trans_dist(x, y)
}
\arguments{
\item{x}{An adjacency matrix}

\item{y}{An adjacency matrix}
}
\value{
Number representing the distance between x and y
}
\description{
Find the transitive distance between two graphs as the number of directed paths that differ between graphs x and y
}
